requirements.txt
python-socketio
python-socketio[client]
eventlet

pytorch
numpy
